Team-10/Backend/requirements.txt
streamlit==1.26.0
langchain==0.0.332
langchain-openai==0.0.2.post1
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.0
imaplib2==3.06
email==4.0.2
python-dateutil==2.8.2 